Extract peak free tree module
